requirements.txt
faiss-cpu
sentence-transformers
transformers
numpy
loguru
pandas
matplotlib
